+++ src/infra/http/dtos/mod.rs
@@ -1,3 +1,4 @@
+pub mod announcements;
 pub mod change_password;
 pub mod comment_on_article;
 pub mod create_article;